split storage key and encryption key
The .NET port did not convert the storage of encrypted passwords correctly.  It was using the encryption key as the dictionary key, storing the encrypted password and encrytion key in the same location.  This makes the encrypted password decryptable if someone were to compromise the underlying data store.  This commit creates a separate storage key for lookup purposes and does not store the encryption key.  A separate token is created that includes the storage key and the encryption key that is only used to generate the URL.  See https://github.com/pinterest/snappass/issues/63 for additional conversation on the original source repo.