Fix preview mechanism
The .NET port included the password on a hidden form on the initial page request.  A button to reveal the password only unhid the form.  This caused the initial page request to expire the password, even if it wasn't viewed.  The intention of the preview feature was to prevent bots that prefetch the URL from destroying the secret.  This commit removes the password from the preview page and adds a new page where the secret is revealed.  Now the secret is only destroyed if the secret is revealed.  Ported based on https://github.com/pinterest/snappass/pull/100